requirements.txt
altair==4.1.0
anyio==3.3.3
appdirs==1.4.4
argh==0.26.2
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
base58==2.1.0
black==21.9b0
bleach==4.1.0
blinker==1.4
boto3==1.18.60
botocore==1.21.60
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.8
cycler==0.10.0
dataclasses==0.8
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
editdistance-s==1.0.0
entrypoints==0.3
filelock==3.3.0
gitdb==4.0.7
GitPython==3.1.24
identify==2.3.0
idna==3.1
importlib-metadata==4.8.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
json5==0.9.5
jsonschema==4.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyter-server==1.11.1
jupyterlab==3.1.18
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
mypy-extensions==0.4.3
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
nodeenv==1.6.0
notebook==6.4.4
numpy==1.21.2
olefile==0.46
packaging==21.0
pandas==1.2.5
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.3.2
pip==21.2.4
platformdirs==2.3.0
pre-commit==2.15.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
pyarrow==5.0.0
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==301
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.3.0
regex==2021.10.8
requests==2.26.0
requests-unixsocket==0.2.0
s3transfer==0.5.0
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
smmap==3.0.5
sniffio==1.2.0
streamlit==1.0.0
terminado==0.9.4
testpath==0.5.0
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.7
validators==0.18.2
virtualenv==20.4.7
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
wheel==0.37.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.6.0
